Put HumanFilesize into it's own component
