+++ package.json
@@ -109,6 +109,7 @@
     "npm": "^6.1.0",
     "nuka-carousel": "^2.3.0",
     "ooni-components": "0.1.4",
+    "prop-types": "^15.6.2",
     "raven-js": "^3.25.2",
     "react-animations": "^0.1.0",
     "react-icons": "^2.2.5",
